Revert "Graduation rehearsal: ToggleButton moves labs → react intact"
This reverts commit 0705a2fdc0a81ca657b45dc44bd119baeef09794.
