--- .changes/graduate-toggle-button.md
@@ -1,10 +1,0 @@
-
----
-core: minor
-labs: major
----
-
-`ToggleButton` graduates from `@urban-ui/labs` to `@urban-ui/react` with an
-unchanged API — import it from `@urban-ui/react/toggle-button`. The labs
-export is removed.
-
