+++ labs/src/index.ts
@@ -1,2 +1,1 @@
-// Emptied by the toggle-button graduation rehearsal — the next experiment exports here.
-export {};
+export { ToggleButton, type ToggleButtonProps } from "./toggle-button/index.js";
