+++ labs/src/toggle-button/__screenshots__/toggle-button.visual/Selected.png
