+++ labs/src/toggle-button/toggle-button.visual.tsx
@@ -1,4 +1,4 @@
-import { ToggleButton } from "@urban-ui/react/toggle-button";
+import { ToggleButton } from "@urban-ui/labs/toggle-button";
 
 /**
  * Visual scenes: coverage, not pedagogy — one export = one screenshot.
