+++ packages/react/package.json
@@ -30,11 +30,6 @@
       "types": "./dist/button/index.d.ts",
       "default": "./dist/button/index.js"
     },
-    "./toggle-button": {
-      "source": "./src/toggle-button/index.ts",
-      "types": "./dist/toggle-button/index.d.ts",
-      "default": "./dist/toggle-button/index.js"
-    },
     "./package.json": "./package.json"
   },
   "scripts": {
