+++ packages/react/src/index.ts
@@ -1,2 +1,1 @@
 export { Button, type ButtonProps } from "./button/index.js";
-export { ToggleButton, type ToggleButtonProps } from "./toggle-button/index.js";
